Support the eth_maxPriorityFeePerGas method
