+++ src/web3-adapter/eth_maxPriorityFeePerGas.ts
@@ -1,0 +1,7 @@
+export function patchEthMaxPriorityFeePerGasMethod(web3: any): void {
+  web3.eth.customRPC({
+    name: "getMaxPriorityFeePerGas",
+    call: "eth_maxPriorityFeePerGas",
+    params: 0,
+  });
+}
